Update Comment.js, IssueDetail.js : Add Comment 부분 CommentList.js로 병합 후 이름 Comment.js로 변경
